updated/populated tables
had to make a couple of changes to the current tables and cleaned it up a bit. Each table should have about ~3 tuples. We may need to visit our current queries to account for these changes. Feel free to look over the changes and let me know if you think something needs to be tweaked!